+++ src/views/Transactions/TransactionsView.js
@@ -1,8 +1,8 @@
 import Observer from "../../utils/observers/Observer.js";
-import monthStore from "../../stores/MonthStore.js";
+import monthState from "../../stores/subjects/MonthState.js";
 import DailyHistory from "../../components/DailyHistory/DailyHistory.js";
 
-const { year, month } = monthStore.getMonthInfo();
+const { year, month } = monthState.getMonthInfo();
 
 const getSortedDailyData = (transactions, filterState) => {
   const filteredTransactions = transactions.filter((transaction) => {
